Extract components from explore.
