+++ app/containers/ExplorePage.tsx
@@ -1,24 +1,15 @@
 import React, { useState } from 'react';
-import Plot from 'react-plotly.js';
 import { useTranslation } from 'react-i18next';
 import { csv } from 'd3-fetch';
-import { Column, Table, Cell } from '@blueprintjs/table';
 import { Card, Elevation } from '@blueprintjs/core';
-import _ from 'lodash';
 import Map from '../components/Map';
+import AnimalsPlot from '../components/AnimalsPlot';
+import ObservationsTable from '../components/ObservationsTable';
 
-type Observation = {
-  exif_datetime: string;
-  pred_1: string;
-  score_1: string;
-  station: string;
-  check: string;
-  camera: string;
-  exif_gps_lat: string;
-};
-
-// eslint-disable-next-line
-function chooseFile(changeFileChoice: (path: string) => void, setData: (data: Observation[]) => void) {
+function chooseFile(
+  changeFileChoice: (path: string) => void,
+  setData: (data: ObservationsData) => void
+) {
   // eslint-disable-next-line global-require
   const { dialog } = require('electron').remote;
   dialog
@@ -45,7 +36,7 @@ function chooseFile(changeFileChoice: (path: string) => void, setData: (data: Ob
     })
     .then(data => {
       if (data !== undefined) {
-        setData((data as unknown) as Observation[]);
+        setData({ observations: (data as unknown) as Observation[] });
         return data;
       }
       return undefined;
@@ -56,123 +47,34 @@ function chooseFile(changeFileChoice: (path: string) => void, setData: (data: Ob
     });
 }
 
-function getDateFromRow(row: Observation) {
-  const onlyDay = row.exif_datetime.substr(0, row.exif_datetime.indexOf(' '));
-  return _.replace(onlyDay, /:/g, '-');
-}
-
-function preparePlotTrace(animalData: Observation[]) {
-  const animalByDate = _.groupBy(animalData, getDateFromRow);
-  return {
-    x: _.keys(animalByDate),
-    y: _.map(animalByDate, 'length'),
-    type: 'scatter',
-    name: animalData[1].pred_1
-  };
-}
-
-function prepareDataForPlot(data: Observation[]) {
-  const dataSorted = _.sortBy(data, 'exif_datetime');
-  const groupAnimal = _.groupBy(dataSorted, 'pred_1');
-  return _.map(groupAnimal, preparePlotTrace);
-}
-
-const selectorOptions = {
-  buttons: [
-    {
-      step: 'month',
-      stepmode: 'backward',
-      count: 1,
-      label: '1m'
-    },
-    {
-      step: 'month',
-      stepmode: 'backward',
-      count: 6,
-      label: '6m'
-    },
-    {
-      step: 'year',
-      stepmode: 'todate',
-      count: 1,
-      label: 'YTD'
-    },
-    {
-      step: 'year',
-      stepmode: 'backward',
-      count: 1,
-      label: '1y'
-    },
-    {
-      step: 'all'
-    }
-  ]
-};
-
 export default function ExplorePage() {
   const { t } = useTranslation();
   const [filePath, setFilePath] = useState<string>();
-  const [data, setData] = useState<undefined | Observation[]>();
+  const [data, setData] = useState<undefined | ObservationsData>();
 
-  const dataForPlot = data !== undefined ? prepareDataForPlot(data) : undefined;
-
-  const table = data && (
-    <Table numRows={data.length} enableColumnReordering>
-      <Column
-        name="Station"
-        cellRenderer={(rowIndex: number) => (
-          <Cell>{data[rowIndex].station}</Cell>
-        )}
-      />
-      <Column
-        name="Check"
-        cellRenderer={(rowIndex: number) => <Cell>{data[rowIndex].check}</Cell>}
-      />
-      <Column
-        name="Camera"
-        cellRenderer={(rowIndex: number) => (
-          <Cell>{data[rowIndex].camera}</Cell>
-        )}
-      />
-      <Column
-        name="Date and time"
-        cellRenderer={(rowIndex: number) => (
-          <Cell>{data[rowIndex].exif_datetime}</Cell>
-        )}
-      />
-      <Column
-        name="GPS"
-        cellRenderer={(rowIndex: number) => (
-          <Cell>{data[rowIndex].exif_gps_lat}</Cell>
-        )}
-      />
-      <Column
-        name="Predicted class"
-        cellRenderer={(rowIndex: number) => (
-          <Cell>{data[rowIndex].pred_1}</Cell>
-        )}
-      />
-      <Column
-        name="Certainty"
-        cellRenderer={(rowIndex: number) => (
-          <Cell>{data[rowIndex].score_1}</Cell>
-        )}
-      />
-    </Table>
-  );
-
-  const layout = {
-    responsive: true,
-    title: 'Animals count over time',
-    hovermode: 'closest',
-    xaxis: {
-      rangeselector: selectorOptions,
-      rangeslider: {}
-    }
-  };
-  return (
-    <div style={{ padding: '10px 30px', width: '100%', overflowY: 'scroll' }}>
-      <h1>{t('Explore')}</h1>
+  const contents =
+    data !== undefined ? (
+      <>
+        <div style={{ flex: 1, paddingBottom: '20px' }}>
+          <Card interactive elevation={Elevation.TWO}>
+            <h2 style={{ marginTop: 0 }}>{t('Filters')}</h2>
+          </Card>
+        </div>
+        ,
+        <div style={{ flex: 1, width: '100%', paddingBottom: '20px' }}>
+          <Card interactive elevation={Elevation.TWO}>
+            <AnimalsPlot data={data} />
+          </Card>
+        </div>
+        ,
+        <div style={{ flex: 1, width: '100%', paddingBottom: '20px' }}>
+          <Card interactive elevation={Elevation.TWO}>
+            <Map data={data} />
+          </Card>
+        </div>
+        <ObservationsTable data={data} />
+      </>
+    ) : (
       <div
         className="bp3-input-group"
         style={{ marginBottom: '10px', width: '60%' }}
@@ -195,31 +97,12 @@ export default function ExplorePage() {
           }}
         />
       </div>
-      {filePath && (
-        <div style={{ flex: 1, paddingBottom: '20px' }}>
-          <Card interactive elevation={Elevation.TWO}>
-            <h2 style={{ marginTop: 0 }}>{t('Filters')}</h2>
-          </Card>
-        </div>
-      )}
-      {filePath && (
-        <div style={{ flex: 1, width: '100%', paddingBottom: '20px' }}>
-          <Card interactive elevation={Elevation.TWO}>
-            <Plot
-              data={dataForPlot as any}
-              useResizeHandler
-              style={{ width: '100%' }}
-              layout={layout as any}
-            />
-          </Card>
-        </div>
-      )}
-      <div style={{ flex: 1, width: '100%', paddingBottom: '20px' }}>
-        <Card interactive elevation={Elevation.TWO}>
-          <Map />
-        </Card>
-      </div>
-      {table}
+    );
+
+  return (
+    <div style={{ padding: '10px 30px', width: '100%', overflowY: 'scroll' }}>
+      <h1>{t('Explore')}</h1>
+      {contents}
     </div>
   );
 }
